FaceAttendance/requirements.txt
numpy==2.0.0
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
pillow==10.3.0
pygame==2.6.0
openpyxl==3.1.4


